+++ src/templates/category.jsx
@@ -7,11 +7,11 @@ import styled from 'styled-components';
 import { useIntl } from 'react-intl';
 
 import AuthorName from '../components/AuthorName';
-import Button from '../components/Button';
 import CategoryName from '../components/CategoryName';
 import Intl from '../components/Intl';
 import Layout from '../components/layout/Layout';
 import LocalizedLink from '../components/link/LocalizedLink';
+import LocalizedLinkButton from '../components/link/LocalizedLinkButton';
 import Rule from '../components/Rule';
 import Theme from '../components/Theme';
 
@@ -135,9 +135,9 @@ const Page = (props) => {
               <p>{post.body.childMarkdownRemark.excerpt}</p>
             </PostContent>
             <Center>
-              <Button to={post.contentful_id} locale={props.pageContext.locale}>
+              <LocalizedLinkButton to={post.contentful_id} locale={props.pageContext.locale}>
                 {i('blogReadMore')}
-              </Button>
+              </LocalizedLinkButton>
             </Center>
             <Separator>
               <Rule color="blue" mode={2} />